src/requirements.txt
rpi.gpio~=0.7.1
